Release version 2.10 revision 3
Using tarball (with SHA1):
2cee6fa50837bfd9f74454460af39a1b91ad46af *LidarViewer-2.10.tar.gz

+++ LidarOctree.cpp
@@ -1073,7 +1073,7 @@ std::string getLidarPartFileName(const char* lidarFileName,const char* partFileN
 
 }
 
-LidarOctree::LidarOctree(const char* lidarFileName,unsigned int sCacheSize,unsigned int sGlCacheSize)
+LidarOctree::LidarOctree(const char* lidarFileName,size_t sCacheSize,size_t sGlCacheSize)
 	:indexFile(getLidarPartFileName(lidarFileName,"Index").c_str()),
 	 pointsFile(getLidarPartFileName(lidarFileName,"Points").c_str()),
 	 normalsFile(0),
